feat: allow custom pickers
