+++ lua/xcodebuild/init.lua
@@ -262,6 +262,7 @@ function M.setup(options)
   local oilNvim = require("xcodebuild.integrations.oil-nvim")
   local neoTree = require("xcodebuild.integrations.neo-tree")
   local xcodeBuildServer = require("xcodebuild.integrations.xcode-build-server")
+  local pickers = require("xcodebuild.ui.pickers")
 
   autocmd.setup()
   projectConfig.load_settings()
@@ -274,6 +275,7 @@ function M.setup(options)
   oilNvim.setup()
   neoTree.setup()
   xcodeBuildServer.setup()
+  pickers.setup()
   setupHighlights()
   warnAboutOldConfig()
