+++ scripts/help-update.sh
@@ -52,8 +52,10 @@ lemmy-help \
   ./lua/xcodebuild/integrations/quick.lua \
   ./lua/xcodebuild/integrations/xcode-build-server.lua \
   ./lua/xcodebuild/integrations/xcodebuild-offline.lua \
+  ./lua/xcodebuild/integrations/telescope.lua \
   ./lua/xcodebuild/ui/pickers.lua \
   ./lua/xcodebuild/ui/picker_actions.lua \
+  ./lua/xcodebuild/ui/pickers_utils.lua \
   ./lua/xcodebuild/helpers.lua \
   ./lua/xcodebuild/util.lua \
   >doc/xcodebuild.txt
